add max, and fix tests
