+++ library/src/app/components/trade/index.ts
@@ -1,0 +1,4 @@
+export { TradeComponent } from './trade.component';
+export { TradeBalanceComponent } from './trade-balance/trade-balance.component';
+export { TradeConfirmComponent } from './trade-confirm/trade-confirm.component';
+export { TradeSummaryComponent } from './trade-summary/trade-summary.component';
