+++ library/src/app/components/trade/trade-balance/trade-balance.component.scss
