+++ library/src/app/components/trade/trade-confirm/trade-confirm.component.ts
