+++ library/src/app/components/trade/trade-summary/trade-summary.component.html
