+++ library/src/app/components/trade/trade.component.scss
@@ -5,7 +5,7 @@
 }
 
 .cybrid-prefix {
-  margin-top: 4px;
+  margin-top: 5px;
   position: relative;
   .cybrid-amount-divider {
     position: absolute;
@@ -18,14 +18,19 @@
 }
 
 .cybrid-option {
+  position: relative;
   display: flex;
   justify-content: flex-start;
   align-items: center;
   gap: 0.5rem;
 }
 
+.cybrid-trigger-name {
+  margin-left: 2.2rem;
+}
+
 .cybrid-trigger-icon {
-  position: fixed;
+  position: absolute;
   max-height: 1.7rem;
   max-width: 1.7rem;
 }
@@ -35,20 +40,22 @@
   max-width: 1.7rem;
 }
 
-.cybrid-trigger-name {
-  margin-left: 2.2rem;
-}
-
 .cybrid-amount-input {
   margin-left: 15px;
 }
 
 .cybrid-approx {
-  margin-left: 0.65rem;
+  margin: 0 0.65rem;
   display: flex;
   align-items: center;
   gap: 0.2rem;
   flex-basis: 50%;
+}
+
+.cybrid-max {
+  line-height: 0;
+  padding: 0;
+  min-width: 0;
 }
 
 @media screen and (min-width: 1535px) {
