+++ library/src/app/modules/routing.module.ts
@@ -13,7 +13,6 @@ import {
 } from '@components';
 
 import { ComponentGuard } from '@guards';
-import { TestTradeComponent } from '../components/test-trade/test-trade/test-trade.component';
 
 export const routes: Routes = [
   {
@@ -62,10 +61,6 @@ export const routes: Routes = [
         path: 'transfer',
         component: TransferComponent,
         canActivate: [ComponentGuard]
-      },
-      {
-        path: 'test-trade',
-        component: TestTradeComponent
       }
     ]
   },
