+++ library/src/shared/services/index.ts
@@ -8,3 +8,4 @@ export { AccountService, Account } from './account/account.service';
 export { RoutingService, RoutingData } from './routing/routing.service';
 export { IdentityVerificationService } from './identity-verification/identity-verification.service';
 export { BankAccountService } from './bank-account/bank-account.service';
+export { PriceService } from './price/price.service';
