+++ library/src/shared/services/price/price.service.spec.ts
@@ -1,0 +1,21 @@
+import { TestBed } from '@angular/core/testing';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+
+// Services
+import { PriceService } from '@services';
+
+describe('PriceService', () => {
+  let service: PriceService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [BrowserAnimationsModule, HttpClientTestingModule]
+    });
+    service = TestBed.inject(PriceService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
