--- library/src/shared/services/test-accounts/test-accounts.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { TestAccountsService } from './test-accounts.service';
-
-describe('TestAccountsService', () => {
-  let service: TestAccountsService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(TestAccountsService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
